requirements.txt
annotated-types==0.7.0
attrs==25.4.0
Automat==25.4.16
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.20.1
greenlet==3.3.0
h11==0.16.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==6.0.2
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyDispatcher==2.0.7
pyOpenSSL==25.3.0
PySocks==1.7.1
python-dotenv==1.2.1
queuelib==1.8.0
requests==2.32.5
requests-file==3.0.1
Scrapy==2.13.4
selenium==4.39.0
service-identity==24.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
tldextract==5.3.0
trio==0.32.0
trio-websocket==0.12.2
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
w3lib==2.3.1
websocket-client==1.9.0
wsproto==1.3.2
zope.interface==8.1.1
